+++ lib/puppet/type/wls_jms_queue.rb
@@ -58,7 +58,7 @@ module Puppet
     property :messagelogging
     property :destination_keys
     property :insertionpaused
-    property :consumtionpaused
+    property :consumptionpaused
     property :productionpaused
 
     add_title_attributes(:jmsmodule, :queue_name) do