+++ lib/puppet/type/wls_jms_topic.rb
@@ -58,7 +58,7 @@ module Puppet
     property :forwardingpolicy
     property :destination_keys
     property :insertionpaused
-    property :consumtionpaused
+    property :consumptionpaused
     property :productionpaused
 
     add_title_attributes(:jmsmodule, :topic_name) do